# myfoodapp/MyFoodApp.API/openapi.yaml
swagger: '2.0'
info:
  title: MyFoodApp.API
  version: '1.0'
host: api-dot-aqueous-aileron-281300.appspot.com
...
paths:
...
  '/api/item/order/{cryptOrderId}':
    get:
      tags:
        - Item
      operationId: GetItemsByOrderId
      produces:
        - text/plain
        - application/json
        - text/json
      parameters:
        - in: path
          name: cryptOrderId
          required: true
          type: string
      responses:
        '200':
          description: Success
          schema:
            type: array
            items:
              $ref: '#/definitions/Item'
...
  '/api/option/order/{cryptOrderId}':
    get:
      tags:
        - Option
      operationId: GetOptionByOrderId
      produces:
        - text/plain
        - application/json
        - text/json
      parameters:
        - in: path
          name: cryptOrderId
          required: true
          type: string
      responses:
        '200':
          description: Success
          schema:
            type: array
            items:
              $ref: '#/definitions/Option'
...
  '/api/order/{orderCryptId}/{newStatusCrypt}':
    patch:
      tags:
        - Order
      operationId: UpdateEncryptedOrderStatus
      produces:
        - text/plain
        - application/json
        - text/json
      parameters:
        - in: path
          name: orderCryptId
          required: true
          type: string
        - in: path
          name: newStatusCrypt
          required: true
          type: string
      responses:
        '200':
          description: Success
          schema:
            $ref: '#/definitions/OrderDto'
  '/api/order/{cryptId}':
    get:
      tags:
        - Order
      operationId: GetByOrderId
      produces:
        - text/plain
        - application/json
        - text/json
      parameters:
        - in: path
          name: cryptId
          required: true
          type: string
      responses:
        '200':
          description: Success
          schema:
            $ref: '#/definitions/OrderDto'
...
  '/api/orderextra/order/{cryptOrderId}':
    get:
      tags:
        - OrderExtra
      operationId: GetAllByOrderId
      parameters:
        - in: path
          name: cryptOrderId
          required: true
          type: string
      responses:
        '200':
          description: Success
...
  '/api/orderitem/order/{cryptOrderId}':
    get:
      tags:
        - OrderItem
      operationId: GetOrderItemsByOrderId
      parameters:
        - in: path
          name: cryptOrderId
          required: true
          type: string
      responses:
        '200':
          description: Success
...
  '/api/orderitemside/order/{cryptOrderId}':
    get:
      tags:
        - OrderItemSide
      operationId: GetItemSideByOrderId
      produces:
        - text/plain
        - application/json
        - text/json
      parameters:
        - in: path
          name: cryptOrderId
          required: true
          type: string
      responses:
        '200':
          description: Success
          schema:
            type: array
            items:
              $ref: '#/definitions/OrderItemSelectedSides'
...
  '/api/side/order/{cryptOrderId}':
    get:
      tags:
        - Side
      operationId: GetSideByOrderId
      produces:
        - text/plain
        - application/json
        - text/json
      parameters:
        - in: path
          name: cryptOrderId
          required: true
          type: string
      responses:
        '200':
          description: Success
          schema:
            type: array
            items:
              $ref: '#/definitions/Side'
...
definitions:
...
  GenderEnum:
    format: int32
    enum:
      - 1
      - 2
    type: integer
  AppUserModel:
    type: object
    properties:
      id:
        format: int64
        type: integer
      firstName:
        type: string
      lastName:
        type: string
      phoneNumber:
        type: string
      email:
        type: string
      userName:
        type: string
      preferences:
        type: string
      dob:
        format: date-time
        type: string
      gender:
        $ref: '#/definitions/GenderEnum'
...
  PaymentProvider:
    type: object
    properties:
      id:
        format: int32
        type: integer
      name:
        type: string
      contact:
        type: string
      email:
        type: string
      phone:
        type: string
      address:
        type: string
      website:
        type: string
      cityId:
        format: int32
        type: integer
      countryId:
        format: int32
        type: integer
      paymentTypes:
        type: array
        items:
          $ref: '#/definitions/PaymentType'
  BranchesPaymentTypes:
    type: object
    properties:
      branchId:
        format: int32
        type: integer
      branch:
        $ref: '#/definitions/Branch'
      paymentTypeId:
        format: int32
        type: integer
      paymentType:
        $ref: '#/definitions/PaymentType'
...
  Reservation:
    type: object
    properties:
      id:
        format: int64
        type: integer
      branchId:
        format: int32
        type: integer
      branch:
        $ref: '#/definitions/Branch'
      userId:
        format: int64
        type: integer
      user:
        $ref: '#/definitions/User'
      reservationDateTime:
        format: date-time
        type: string
      requestedDateTime:
        format: date-time
        type: string
      numberOfGuests:
        format: int32
        type: integer
      specialEventId:
        format: int64
        type: integer
      currentStatusId:
        format: int32
        type: integer
      eventId:
        format: int64
        type: integer
      event:
        $ref: '#/definitions/Event'
      transactionId:
        format: int64
        type: integer
      transaction:
        $ref: '#/definitions/ClientTransaction'
      specialEvent:
        $ref: '#/definitions/ReservationSpecialEvent'
      reservationStatus:
        $ref: '#/definitions/ReservationStatus'
      reservationStatusHistories:
        type: array
        items:
          $ref: '#/definitions/ReservationStatusHistory'
      notes:
        type: string
  BranchesEventTypes:
    type: object
    properties:
      branchId:
        format: int32
        type: integer
      branch:
        $ref: '#/definitions/Branch'
      eventTypeId:
        format: int32
        type: integer
      eventType:
        $ref: '#/definitions/EventType'
  EventType:
    type: object
    properties:
      id:
        format: int32
        type: integer
      name:
        type: string
      deliverable:
        type: boolean
      events:
        type: array
        items:
          $ref: '#/definitions/Event'
      branchesEventTypes:
        type: array
        items:
          $ref: '#/definitions/BranchesEventTypes'
...
  Coupon:
    type: object
    properties:
      id:
        format: int64
        type: integer
      name:
        type: string
      discountType:
        format: int32
        type: integer
      type:
        format: int32
        type: integer
      minAmount:
        format: double
        type: number
      amount:
        format: double
        type: number
      code:
        type: string
      expirationType:
        format: int32
        type: integer
      endDate:
        format: date-time
        type: string
      limit:
        format: int32
        type: integer
      isActive:
        type: boolean
      branchId:
        format: int32
        type: integer
      clientId:
        format: int32
        type: integer
      client:
        $ref: '#/definitions/Client'
      branch:
        $ref: '#/definitions/Branch'
      events:
        type: array
        items:
          $ref: '#/definitions/Event'
  OrderStatus:
    type: object
    properties:
      id:
        format: int32
        type: integer
      name:
        type: string
      billable:
        type: boolean
      orders:
        type: array
        items:
          $ref: '#/definitions/Order'
  OrderStatusHistory:
    type: object
    properties:
      id:
        format: int64
        type: integer
      orderId:
        format: int64
        type: integer
      statusId:
        format: int32
        type: integer
      changeDateTime:
        format: date-time
        type: string
      adminUserId:
        format: int64
        type: integer
      adminUser:
        $ref: '#/definitions/User'
      order:
        $ref: '#/definitions/Order'
  Menu:
    type: object
    properties:
      id:
        format: int64
        type: integer
      branchId:
        format: int32
        type: integer
      name:
        type: string
      isActive:
        type: boolean
      branch:
        $ref: '#/definitions/Branch'
      menusCategories:
        type: array
        items:
          $ref: '#/definitions/MenusCategories'
  MenusCategories:
    type: object
    properties:
      menuId:
        format: int64
        type: integer
      menu:
        $ref: '#/definitions/Menu'
      categoryId:
        format: int64
        type: integer
      category:
        $ref: '#/definitions/Category'
  Category:
    type: object
    properties:
      id:
        format: int64
        type: integer
      name:
        type: string
      image:
        type: string
      position:
        format: int32
        type: integer
      isActive:
        type: boolean
      menusCategories:
        type: array
        items:
          $ref: '#/definitions/MenusCategories'
      categoriesItems:
        type: array
        items:
          $ref: '#/definitions/CategoriesItems'
      isVisibleInMenu:
        type: boolean
      branchId:
        format: int32
        type: integer
      branch:
        $ref: '#/definitions/Branch'
  CategoriesItems:
    type: object
    properties:
      itemId:
        format: int64
        type: integer
      item:
        $ref: '#/definitions/Item'
      categoryId:
        format: int64
        type: integer
      category:
        $ref: '#/definitions/Category'
  OrderExtra:
    type: object
    properties:
      id:
        format: int64
        type: integer
      orderId:
        format: int64
        type: integer
      sideId:
        format: int64
        type: integer
      quantity:
        format: int32
        type: integer
      price:
        format: double
        type: number
      side:
        $ref: '#/definitions/Side'
      order:
        $ref: '#/definitions/Order'
  SidePriceHistory:
    type: object
    properties:
      id:
        format: int64
        type: integer
      price:
        format: double
        type: number
      startDate:
        format: date-time
        type: string
      endDate:
        format: date-time
        type: string
      sideId:
        format: int64
        type: integer
      adminUserId:
        format: int64
        type: integer
      side:
        $ref: '#/definitions/Side'
  OrderItemSelectedSides:
    type: object
    properties:
      id:
        format: int64
        type: integer
      itemId:
        format: int64
        type: integer
      orderItemId:
        format: int64
        type: integer
      orderItem:
        $ref: '#/definitions/OrderItem'
      sideId:
        format: int64
        type: integer
      side:
        $ref: '#/definitions/Side'
      price:
        format: double
        type: number
        readOnly: true
  Side:
    type: object
    properties:
      id:
        format: int64
        type: integer
      name:
        type: string
      description:
        type: string
      currentPrice:
        format: double
        type: number
      isActive:
        type: boolean
      image:
        type: string
      itemsSides:
        type: array
        items:
          $ref: '#/definitions/ItemsSides'
      orderExtras:
        type: array
        items:
          $ref: '#/definitions/OrderExtra'
      sidePriceHistories:
        type: array
        items:
          $ref: '#/definitions/SidePriceHistory'
      orderItemSelectedSides:
        type: array
        items:
          $ref: '#/definitions/OrderItemSelectedSides'
      branchId:
        format: int32
        type: integer
      branch:
        $ref: '#/definitions/Branch'
  ItemsSides:
    type: object
    properties:
      itemId:
        format: int64
        type: integer
      item:
        $ref: '#/definitions/Item'
      sideId:
        format: int64
        type: integer
      side:
        $ref: '#/definitions/Side'
  OrderItemSelectedOptions:
    type: object
    properties:
      id:
        format: int64
        type: integer
      orderItemId:
        format: int64
        type: integer
      orderItem:
        $ref: '#/definitions/OrderItem'
      optionId:
        format: int64
        type: integer
      option:
        $ref: '#/definitions/Option'
  Option:
    type: object
    properties:
      id:
        format: int64
        type: integer
      name:
        type: string
      isActive:
        type: boolean
      choices:
        type: string
      itemsOptions:
        type: array
        items:
          $ref: '#/definitions/ItemsOptions'
      orderItemSelectedOptions:
        type: array
        items:
          $ref: '#/definitions/OrderItemSelectedOptions'
      branchId:
        format: int32
        type: integer
      branch:
        $ref: '#/definitions/Branch'
  ItemsOptions:
    type: object
    properties:
      itemId:
        format: int64
        type: integer
      item:
        $ref: '#/definitions/Item'
      optionId:
        format: int64
        type: integer
      option:
        $ref: '#/definitions/Option'
  ItemPriceHistory:
    type: object
    properties:
      id:
        format: int64
        type: integer
      price:
        format: double
        type: number
      startDate:
        format: date-time
        type: string
      endDate:
        format: date-time
        type: string
      itemId:
        format: int64
        type: integer
      adminUserId:
        format: int64
        type: integer
      item:
        $ref: '#/definitions/Item'
      adminUser:
        $ref: '#/definitions/User'
  Item:
    type: object
    properties:
      id:
        format: int64
        type: integer
      name:
        type: string
      description:
        type: string
      position:
        format: int32
        type: integer
      image:
        type: string
      prepTimeMins:
        format: int32
        type: integer
      currentPrice:
        format: double
        type: number
      numberSidesRequired:
        format: int32
        type: integer
      optionsRequired:
        type: boolean
      isActive:
        type: boolean
      categoriesItems:
        type: array
        items:
          $ref: '#/definitions/CategoriesItems'
      itemsSides:
        type: array
        items:
          $ref: '#/definitions/ItemsSides'
      itemsOptions:
        type: array
        items:
          $ref: '#/definitions/ItemsOptions'
      itemPriceHistories:
        type: array
        items:
          $ref: '#/definitions/ItemPriceHistory'
      orderItems:
        type: array
        items:
          $ref: '#/definitions/OrderItem'
      isVisibleInMenu:
        type: boolean
      branchId:
        format: int32
        type: integer
      branch:
        $ref: '#/definitions/Branch'
  OrderItem:
    type: object
    properties:
      id:
        format: int64
        type: integer
      orderId:
        format: int64
        type: integer
      itemId:
        format: int64
        type: integer
      quantity:
        format: int32
        type: integer
      price:
        format: double
        type: number
      notes:
        type: string
      item:
        $ref: '#/definitions/Item'
      order:
        $ref: '#/definitions/Order'
      selectedSides:
        type: array
        items:
          $ref: '#/definitions/OrderItemSelectedSides'
      selectedOptions:
        type: array
        items:
          $ref: '#/definitions/OrderItemSelectedOptions'
  Order:
    type: object
    properties:
      id:
        format: int64
        type: integer
      createOrderDateTime:
        format: date-time
        type: string
      currentStatusId:
        format: int32
        type: integer
      scheduledOrderReady:
        format: date-time
        type: string
      clientOrderId:
        type: string
      eventId:
        format: int64
        type: integer
      event:
        $ref: '#/definitions/Event'
      orderStatus:
        $ref: '#/definitions/OrderStatus'
      orderStatusHistories:
        type: array
        items:
          $ref: '#/definitions/OrderStatusHistory'
      orderItems:
        type: array
        items:
          $ref: '#/definitions/OrderItem'
      orderExtras:
        type: array
        items:
          $ref: '#/definitions/OrderExtra'
...
  Event:
    type: object
    properties:
      id:
        format: int64
        type: integer
      startDateTime:
        format: date-time
        type: string
      endDateTime:
        format: date-time
        type: string
      details:
        type: string
      customerName:
        type: string
      customerId:
        type: string
      appUserId:
        format: int64
        type: integer
      branchId:
        format: int32
        type: integer
      typeId:
        format: int32
        type: integer
      currentStatusId:
        format: int32
        type: integer
      tableNumber:
        type: string
      logisticsProviderId:
        format: int32
        type: integer
      totalAmount:
        format: double
        type: number
      pctMyFoodApp:
        format: double
        type: number
      couponCode:
        type: string
      couponId:
        format: int64
        type: integer
      couponDiscountAmount:
        format: double
        type: number
      destinationLocationId:
        format: int64
        type: integer
      deliveryDistanceKm:
        format: double
        type: number
      deliveryCost:
        format: double
        type: number
      transactionId:
        format: int64
        type: integer
      destinationLocation:
        $ref: '#/definitions/Location'
      user:
        $ref: '#/definitions/AppUserModel'
      branch:
        $ref: '#/definitions/Branch'
      reservation:
        $ref: '#/definitions/Reservation'
      eventType:
        $ref: '#/definitions/EventType'
      eventStatus:
        $ref: '#/definitions/EventStatus'
      transaction:
        $ref: '#/definitions/ClientTransaction'
      coupon:
        $ref: '#/definitions/Coupon'
      orders:
        type: array
        items:
          $ref: '#/definitions/Order'
      eventStatusHistories:
        type: array
        items:
          $ref: '#/definitions/EventStatusHistory'
  Country:
    type: object
    properties:
      id:
        format: int32
        type: integer
      countryName:
        type: string
      countryPhoneCode:
        type: string
      cities:
        type: array
        items:
          $ref: '#/definitions/City'
      locations:
        type: array
        items:
          $ref: '#/definitions/Location'
      paymentProviders:
        type: array
        items:
          $ref: '#/definitions/PaymentProvider'
  City:
    type: object
    properties:
      id:
        format: int32
        type: integer
      cityName:
        type: string
      countryId:
        format: int32
        type: integer
      country:
        $ref: '#/definitions/Country'
      locations:
        type: array
        items:
          $ref: '#/definitions/Location'
      paymentProviders:
        type: array
        items:
          $ref: '#/definitions/PaymentProvider'
      zones:
        type: array
        items:
          $ref: '#/definitions/Zone'
  Zone:
    type: object
    properties:
      id:
        format: int32
        type: integer
      zoneName:
        type: string
      coordinates:
        type: string
      cityId:
        format: int32
        type: integer
      city:
        $ref: '#/definitions/City'
      locations:
        type: array
        items:
          $ref: '#/definitions/Location'
      branches:
        type: array
        items:
          $ref: '#/definitions/Branch'
  Location:
    type: object
    properties:
      id:
        format: int64
        type: integer
      coordinates:
        type: string
      address:
        type: string
      zoneId:
        format: int32
        type: integer
      cityId:
        format: int32
        type: integer
      countryId:
        format: int32
        type: integer
      iconIndex:
        type: string
      notes:
        type: string
      eventId:
        format: int64
        type: integer
      event:
        $ref: '#/definitions/Event'
      userId:
        format: int64
        type: integer
      user:
        $ref: '#/definitions/User'
      zone:
        $ref: '#/definitions/Zone'
      city:
        $ref: '#/definitions/City'
      country:
        $ref: '#/definitions/Country'
  Favorite:
    type: object
    properties:
      id:
        format: int64
        type: integer
      menuItemId:
        format: int64
        type: integer
      userId:
        format: int64
        type: integer
      user:
        $ref: '#/definitions/User'
...
  UserNotification:
    type: object
    properties:
      id:
        format: int64
        type: integer
      message:
        type: string
      notificationTypeId:
        format: int32
        type: integer
      isRead:
        type: boolean
      userId:
        format: int64
        type: integer
      user:
        $ref: '#/definitions/User'
      userNotificationType:
        $ref: '#/definitions/UserNotificationType'
...
  UserPointsTransaction:
    type: object
    properties:
      id:
        format: int64
        type: integer
      date:
        format: date-time
        type: string
      userId:
        format: int64
        type: integer
      transactionType:
        $ref: '#/definitions/UserPointTransactionTypeEnum'
      points:
        format: int32
        type: integer
      action:
        $ref: '#/definitions/ActionPointTypeEnum'
      notes:
        type: string
      pointEventId:
        format: int64
        type: integer
      pointsExchangeId:
        format: int32
        type: integer
      user:
        $ref: '#/definitions/User'
      pointsExchange:
        $ref: '#/definitions/PointsExchange'
  User:
    type: object
    properties:
      firstName:
        type: string
      lastName:
        type: string
      dob:
        format: date-time
        type: string
      gender:
        $ref: '#/definitions/GenderEnum'
      preferences:
        type: array
        items:
          type: string
      locations:
        type: array
        items:
          $ref: '#/definitions/Location'
      favorites:
        type: array
        items:
          $ref: '#/definitions/Favorite'
      appUserNotifications:
        type: array
        items:
          $ref: '#/definitions/UserNotification'
      events:
        type: array
        items:
          $ref: '#/definitions/Event'
      orderStatusHistories:
        type: array
        items:
          $ref: '#/definitions/OrderStatusHistory'
      clientsAdminUsers:
        type: array
        items:
          $ref: '#/definitions/ClientsAdminUsers'
      userPointTransactions:
        type: array
        items:
          $ref: '#/definitions/UserPointsTransaction'
      id:
        format: int64
        type: integer
      userName:
        type: string
      normalizedUserName:
        type: string
      email:
        type: string
      normalizedEmail:
        type: string
      emailConfirmed:
        type: boolean
      passwordHash:
        type: string
      securityStamp:
        type: string
      concurrencyStamp:
        type: string
      phoneNumber:
        type: string
      phoneNumberConfirmed:
        type: boolean
      twoFactorEnabled:
        type: boolean
      lockoutEnd:
        format: date-time
        type: string
      lockoutEnabled:
        type: boolean
      accessFailedCount:
        format: int32
        type: integer
  ClientsAdminUsers:
    type: object
    properties:
      clientId:
        format: int32
        type: integer
      client:
        $ref: '#/definitions/Client'
      userId:
        format: int64
        type: integer
      user:
        $ref: '#/definitions/User'
  Client:
    type: object
    properties:
      id:
        format: int32
        type: integer
      clientName:
        type: string
      isActive:
        type: boolean
      bannerURL:
        type: string
      logoURL:
        type: string
      branches:
        type: array
        items:
          $ref: '#/definitions/Branch'
      clientsUsers:
        type: array
        items:
          $ref: '#/definitions/ClientsAdminUsers'
      coupons:
        type: array
        items:
          $ref: '#/definitions/Coupon'
  BranchPreferences:
    type: object
    properties:
      branchId:
        format: int32
        type: integer
      preferenceName:
        type: string
      preferenceValue:
        type: string
      type:
        type: string
      superAdminOnly:
        type: boolean
      branch:
        $ref: '#/definitions/Branch'
  TimeSpan:
    type: object
    properties:
      ticks:
        format: int64
        type: integer
        readOnly: true
      days:
        format: int32
        type: integer
        readOnly: true
      hours:
        format: int32
        type: integer
        readOnly: true
      milliseconds:
        format: int32
        type: integer
        readOnly: true
      minutes:
        format: int32
        type: integer
        readOnly: true
      seconds:
        format: int32
        type: integer
        readOnly: true
      totalDays:
        format: double
        type: number
        readOnly: true
      totalHours:
        format: double
        type: number
        readOnly: true
      totalMilliseconds:
        format: double
        type: number
        readOnly: true
      totalMinutes:
        format: double
        type: number
        readOnly: true
      totalSeconds:
        format: double
        type: number
        readOnly: true
  BranchExceptionDate:
    type: object
    properties:
      id:
        format: int32
        type: integer
      branchId:
        format: int32
        type: integer
      service:
        type: string
      isClosed:
        type: boolean
      branch:
        $ref: '#/definitions/Branch'
      exceptionDate:
        format: date-time
        type: string
      timeStart1:
        $ref: '#/definitions/TimeSpan'
      timeStart2:
        $ref: '#/definitions/TimeSpan'
      timeStart3:
        $ref: '#/definitions/TimeSpan'
      timeEnd1:
        $ref: '#/definitions/TimeSpan'
      timeEnd2:
        $ref: '#/definitions/TimeSpan'
      timeEnd3:
        $ref: '#/definitions/TimeSpan'
...
  BranchesGroups:
    type: object
    properties:
      branchId:
        format: int32
        type: integer
      branch:
        $ref: '#/definitions/Branch'
      groupId:
        format: int32
        type: integer
      group:
        $ref: '#/definitions/Group'
...
  BranchLogisticProvider:
    type: object
    properties:
      branchId:
        format: int32
        type: integer
      branch:
        $ref: '#/definitions/Branch'
      logisticProviderId:
        format: int32
        type: integer
      logisticProvider:
        $ref: '#/definitions/LogisticProvider'
      isBranchDefault:
        type: boolean
  Branch:
    type: object
    properties:
      name:
        type: string
      address:
        type: string
      email:
        type: string
      phone:
        type: string
      mobilePhone:
        type: string
      whatsapp:
        type: string
      website:
        type: string
      nit:
        type: string
      coordinates:
        type: string
      zoneId:
        format: int32
        type: integer
      bannerUrl:
        type: string
      logoUrl:
        type: string
      cityId:
        format: int32
        type: integer
      countryId:
        format: int32
        type: integer
      clientId:
        format: int32
        type: integer
      isActive:
        type: boolean
      tags:
        type: array
        items:
          type: string
      client:
        $ref: '#/definitions/Client'
      menus:
        type: array
        items:
          $ref: '#/definitions/Menu'
      zone:
        $ref: '#/definitions/Zone'
      city:
        $ref: '#/definitions/City'
      country:
        $ref: '#/definitions/Country'
      branchPreferences:
        type: array
        items:
          $ref: '#/definitions/BranchPreferences'
      branchExceptionDates:
        type: array
        items:
          $ref: '#/definitions/BranchExceptionDate'
      events:
        type: array
        items:
          $ref: '#/definitions/Event'
      branchesPaymentTypes:
        type: array
        items:
          $ref: '#/definitions/BranchesPaymentTypes'
      branchesGroups:
        type: array
        items:
          $ref: '#/definitions/BranchesGroups'
      branchesEventTypes:
        type: array
        items:
          $ref: '#/definitions/BranchesEventTypes'
      coupons:
        type: array
        items:
          $ref: '#/definitions/Coupon'
      branchLogisticProviders:
        type: array
        items:
          $ref: '#/definitions/BranchLogisticProvider'
      imageCollection:
        $ref: '#/definitions/ImageCollection'
      id:
        type: string
  ImageCollection:
    type: object
    properties:
      id:
        format: int32
        type: integer
      sectionName:
        $ref: '#/definitions/AdSectionEnum'
      branchId:
        format: int32
        type: integer
      name:
        type: string
      isActive:
        type: boolean
      adImages:
        type: array
        items:
          $ref: '#/definitions/AdImage'
      branch:
        $ref: '#/definitions/Branch'
...
  Int32KeyValue:
    type: object
    properties:
      id:
        format: int32
        type: integer
      description:
        type: string
  OrderItemDto:
    type: object
    properties:
      id:
        format: int32
        type: integer
      orderItemId:
        format: int32
        type: integer
      extraId:
        format: int32
        type: integer
      detailId:
        type: string
        readOnly: true
      quantity:
        format: int32
        type: integer
      name:
        type: string
      description:
        type: string
      prepTimeMins:
        format: int32
        type: integer
      currentPrice:
        format: double
        type: number
      option:
        $ref: '#/definitions/Int32KeyValue'
      sides:
        type: array
        items:
          $ref: '#/definitions/Int32KeyValue'
  AddressDto:
    type: object
    properties:
      street:
        type: string
      number:
        type: string
      depto:
        type: string
      reference:
        type: string
      coordinates:
        type: object
  CustomerDto:
    type: object
    properties:
      address:
        $ref: '#/definitions/AddressDto'
      id:
        type: string
      name:
        type: string
      email:
        type: string
  PaymentDto:
    type: object
    properties:
      transactionImage:
        type: string
      reference:
        type: string
      methodType:
        type: string
      transactionConfirmed:
        type: boolean
  OrderDto:
    type: object
    properties:
      orderDetail:
        type: array
        items:
          $ref: '#/definitions/OrderItemDto'
      customer:
        $ref: '#/definitions/CustomerDto'
      payment:
        $ref: '#/definitions/PaymentDto'
      orderStatus:
        format: int32
        type: integer
      id:
        format: int64
        type: integer
...